created modal for create channel
