+++ react-app/src/components/Main/index.js
@@ -28,7 +28,7 @@ export default function Servers() {
   const [serverActive, setServerActive] = useState(false);
   const [createNewServer, setCreateNewServer] = useState(false);
   const [joinServer, setJoinServer] = useState(false);
-  const [channelModalActive, setChannelModalActive] = useState(false)
+  // const [channelModalActive, setChannelModalActive] = useState(false)
 
   useEffect(() => {
     socket = io();
@@ -78,40 +78,40 @@ export default function Servers() {
     setJoinServer(true);
   }
 
-  function addChannelFunc() {
-    return (
-      <>
-        <div
-          className={style.serverModalBackground}
-          onClick={() => setChannelModalActive(false)}
-        ></div>
-        <div id='channelModal' className={style.channelModalContainer}>
-          <div className={style.serverModalWrapper}>
-              {/* <CreateChannel /> */}
-            <div className={style.title}>Create Text Channel</div>
-            <div
-              className={style.closeModal}
-              onClick={() => setChannelModalActive(false)}
-            >
-              <svg
-                className={style.closeX}
-                aria-hidden="false"
-                width="24"
-                height="24"
-                viewBox="0 0 24 24"
-              >
-                <path d="M18.4 4L12 10.4L5.6 4L4 5.6L10.4 12L4 18.4L5.6 20L12 13.6L18.4 20L20 18.4L13.6 12L20 5.6L18.4 4Z"></path>
-              </svg>
-            </div>
-            <div className={style.subheading}>
-              Your server is where you and your friends hang out. Make it yours
-              and start talking.
-            </div>
-          </div>
-        </div>
-      </>
-    );
-  }
+  // function addChannelFunc() {
+  //   return (
+  //     <>
+  //       <div
+  //         className={style.serverModalBackground}
+  //         onClick={() => setChannelModalActive(false)}
+  //       ></div>
+  //       <div id='channelModal' className={style.channelModalContainer}>
+  //         <div className={style.serverModalWrapper}>
+  //             {/* <CreateChannel /> */}
+  //           <div className={style.title}>Create Text Channel</div>
+  //           <div
+  //             className={style.closeModal}
+  //             onClick={() => setChannelModalActive(false)}
+  //           >
+  //             <svg
+  //               className={style.closeX}
+  //               aria-hidden="false"
+  //               width="24"
+  //               height="24"
+  //               viewBox="0 0 24 24"
+  //             >
+  //               <path d="M18.4 4L12 10.4L5.6 4L4 5.6L10.4 12L4 18.4L5.6 20L12 13.6L18.4 20L20 18.4L13.6 12L20 5.6L18.4 4Z"></path>
+  //             </svg>
+  //           </div>
+  //           <div className={style.subheading}>
+  //             Your server is where you and your friends hang out. Make it yours
+  //             and start talking.
+  //           </div>
+  //         </div>
+  //       </div>
+  //     </>
+  //   );
+  // }
 
   function addServerFunc() {
     return (
@@ -167,7 +167,7 @@ export default function Servers() {
 
   return (
     <main className={style.main}>
-      {channelModalActive && addChannelFunc()}
+      {/* {channelModalActive && addChannelFunc()} */}
       {serverActive && addServerFunc()}
       {createNewServer && (
         <CreateServer
@@ -193,7 +193,7 @@ export default function Servers() {
             exact
             path={["/servers/:serverId", "/servers/:serverId/:channelId"]}
           >
-            <ChannelList setChannelModalActive={setChannelModalActive}/>
+            <ChannelList />
           </Route>
         </Switch>
       </div>
