+++ react-app/src/components/Main/servers.module.css
@@ -58,7 +58,7 @@
   z-index: 101;
 }
 
-.channelModalContainer {
+/* .channelModalContainer {
   position: fixed;
   height: fit-content;
   width: 400px;
@@ -69,7 +69,7 @@
   transform: translate(-50%, -50%);
   border-radius: 4px;
   z-index: 101;
-}
+} */
 
 .serverModalWrapper {
   display: flex;
